requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pillow>=10.0.0
spotipy>=2.24.0
requests>=2.31.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
